New IPython Notebook Added
So this is what our rules set looks like. Antecedents are basically items on the LHS and consequent are items on the RHS. You can now start shortlisting the rules and items and start building your strategies.

It is important to mention here that choosing rules is more than just picking the highest lift value. There are different perspectives to take into account. Let's say for example you had a brick and mortar store. You had a offer running for buy a single loaf of bread and get a jar of peanut better free. The lift value would be high but we have take into account the fact that items might be co occuring because we presented them together.

Your business model also affects the thresholds. Ink manufacturers will have to set a lower threshold values for confidence and lift as compared to a retail store. Now back to our strategy. For our email marketing campaign we would select a few but highly scored rules. For our recommendation system, we will rather have an expansive set of rules hence opt towards lower scores. The score that are borderline threshold can also be used. We can showcase them at the cart page and maybe offer a slight discount to encourage up-selling.